archive/prototypes/agent_prototype_v2/requirements.txt
langgraph==0.0.40
litellm==1.81.13
plotly==5.18.0
pandas==2.1.4
langfuse==2.6.3
python-dotenv==1.0.0
scikit-learn==1.3.2
numpy==1.26.2
sqlalchemy
psycopg2-binary
chainlit
asyncpg